添加最新编译的 premake

+++ DashProject/DashCore/Src/Graphics/ShaderResource.cpp
@@ -89,7 +89,7 @@ namespace Dash
 
 		mCompiledShaderBlob = compiledShaderBlob;
 		mShaderBinary.Data = mCompiledShaderBlob->GetBufferPointer();
-		mShaderBinary.Size = mCompiledShaderBlob->GetBufferSize();
+		mShaderBinary.Size = static_cast<uint32_t>(mCompiledShaderBlob->GetBufferSize());
 
 		mCreationInfo = creationInfo;
